refactor: heart를 post와 합침
