+++ src/main/java/com/theocean/fundering/FunderingApplication.java
@@ -2,8 +2,10 @@ package com.theocean.fundering;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
+@EnableScheduling
 public class FunderingApplication {
 
     public static void main(final String[] args) {
